requirements.txt
oso-cloud
fastapi[all]
kafka-python